requirements.txt
dotenv
pydantic
linkup-sdk
crewai[tools]
